import full ASL 2.0 text
All derivative copies of takora are required to contain the full text of
the ASL 2.0 license (as described in the sentence "You must give any
other recipients of the Work or Derivative Works a copy of this
License"). In order to ease development and distribution of takora, it
makes sense to simply include the full text in the upstream project.

This text file is taken verbatim from
https://www.apache.org/licenses/LICENSE-2.0.txt

Signed-off-by: Ken Dreyer <kdreyer@redhat.com>